feat: Adds some documentation to the new modules
