+++ Sources/DatabaseExtensions/Protocols/TableRepresentable.swift
