+++ Sources/DatabaseRouter/DatabaseExecutor.swift
@@ -1,44 +1,49 @@
+import DatabaseExtensions
 import Dependencies
 import DependenciesMacros
 import Foundation
-import Supabase
-import SupabaseExtensions
+import PostgREST
 
 extension DependencyValues {
 
+  /// Access the database executor that can execute queries on the database.
   public var databaseExecutor: DatabaseExecutor {
     get { self[DatabaseExecutor.self] }
     set { self[DatabaseExecutor.self] = newValue }
   }
 }
 
+/// Execute queries on the database and decode the results.
+///
+///
 @DependencyClient
 public struct DatabaseExecutor {
+  
+  /// The json decoder used to decode responses.
   public var decoder: JSONDecoder = .init()
+  
+  /// Execute commands on the database and return the data.
   public var execute: (PostgrestBuilder) async throws -> Data
+  
+  /// Create a query builder for the given table.
   public var query: (AnyTable) throws -> PostgrestQueryBuilder
 
-  func run(_ builder: PostgrestBuilder) async throws {
-    _ = try await self.execute(builder)
-  }
-
-  @discardableResult
-  func run<A: Decodable>(
-    _ builder: PostgrestBuilder
-  ) async throws -> A {
-    let data = try await self.execute(builder)
-    return try decoder.decode(A.self, from: data)
-  }
-
+  /// Run the table route on the database.
   public func run<T: TableRouter>(_ route: T) async throws {
-    try await route.execute(query: self.query, execute: self.run)
+    _ = try await self.execute(
+      route.build(query: query(T.table), route: route)
+    )
   }
 
+  /// Run the table route on the database.
   @discardableResult
   public func run<A: Decodable, T: TableRouter>(
     _ route: T
   ) async throws -> A {
-    try await route.execute(query: self.query, execute: self.run)
+    let data = try await self.execute(
+      route.build(query: query(T.table), route: route)
+    )
+    return try decoder.decode(A.self, from: data)
   }
 
 }
