+++ Sources/DatabaseRouter/Protocols/TableRouter.swift
@@ -1,29 +1,15 @@
-import Dependencies
-import Supabase
-import SupabaseExtensions
+import DatabaseExtensions
 
+/// Represents a router for a given database table.
+///
+/// Generally modeled as an enum for the routes.
 public protocol TableRouter {
-  var table: AnyTable { get }
-  var builder: QueryBuilder<Self> { get }
+  
+  /// The table that the router operates on.
+  static var table: AnyTable { get }
+  
+  /// The query builder for the table, which handles routing for the table.
+  var build: QueryBuilder<Self> { get }
 }
 
-extension TableRouter {
-  func execute(
-    query: @escaping (AnyTable) throws -> PostgrestQueryBuilder,
-    execute: @escaping (PostgrestBuilder) async throws -> Void
-  ) async throws {
-    return try await execute(
-      builder.build(query(self.table), self)
-    )
-  }
 
-  @discardableResult
-  func execute<A: Decodable>(
-    query: @escaping (AnyTable) throws -> PostgrestQueryBuilder,
-    execute: @escaping (PostgrestBuilder) async throws -> A
-  ) async throws -> A {
-    return try await execute(
-      builder.build(query(self.table), self)
-    )
-  }
-}
