+++ Sources/DatabaseRouter/QueryBuilder.swift
@@ -1,10 +1,18 @@
 import Foundation
-import Supabase
+import PostgREST
 
-public struct QueryBuilder<Table> {
-  let build: (PostgrestQueryBuilder, Table) throws -> PostgrestBuilder
+/// Handles routes for a table, returning the database queries that can be executed for the
+/// given table.
+///
+public struct QueryBuilder<TableRoute> {
+  
+  let build: (PostgrestQueryBuilder, TableRoute) throws -> PostgrestBuilder
 
-  public init(build: @escaping (PostgrestQueryBuilder, Table) throws -> PostgrestBuilder) {
+  public init(build: @escaping (PostgrestQueryBuilder, TableRoute) throws -> PostgrestBuilder) {
     self.build = build
+  }
+  
+  public func callAsFunction(query: PostgrestQueryBuilder, route: TableRoute) throws -> PostgrestBuilder {
+    try build(query, route)
   }
 }
