--- Sources/SupabaseExtensions/Extensions/SupabaseClient+from.swift
@@ -1,7 +1,0 @@
-import Supabase
-
-extension SupabaseClient {
-  public func from(_ table: AnyTable) -> PostgrestQueryBuilder {
-    self.from(table.tableName)
-  }
-}
